src/akg/parsers/document_parser.py
"""
Document parser using LlamaParse for advanced document processing.
"""

import asyncio
import logging
import os
import ssl
from pathlib import Path
from typing import Any, Dict, Optional

from bs4 import BeautifulSoup
from docx import Document as DocxDocument

from ..config import config

logger = logging.getLogger(__name__)

# Disable SSL verification globally for macOS issues
os.environ['PYTHONHTTPSVERIFY'] = '0'
os.environ['CURL_CA_BUNDLE'] = ''

# Try to disable SSL warnings
try:
    import urllib3
    urllib3.disable_warnings(urllib3.exceptions.InsecureRequestWarning)
except:
    pass

# Try to import LlamaParse, but handle gracefully if it fails
try:
    from llama_parse import LlamaParse
    LLAMA_PARSE_AVAILABLE = True
except ImportError as e:
    logger.warning(f"LlamaParse not available: {e}")
    LlamaParse = None
    LLAMA_PARSE_AVAILABLE = False

class DocumentParser:
    """Enhanced document parser using LlamaParse and other libraries."""
    
    def __init__(self):
        self.llama_parser = None
        
        # Initialize LlamaParse if available
        if LLAMA_PARSE_AVAILABLE and LlamaParse:
            try:
                self.llama_parser = LlamaParse(
                    api_key=config.llama_cloud_api_key,
                    result_type="markdown",  # Can be "text" or "markdown"
                    verbose=True
                )
                logger.info("LlamaParse initialized successfully")
            except Exception as e:
                logger.warning(f"Failed to initialize LlamaParse: {e}")
                self.llama_parser = None
        else:
            logger.info("LlamaParse not available, using fallback parsers")
        
    async def parse_document(self, file_path: Path) -> str:
        """Parse document content based on file type."""
        file_ext = file_path.suffix.lower()
        
        try:
            if file_ext == '.pdf':
                return await self._parse_pdf(file_path)
            elif file_ext == '.docx':
                return await self._parse_docx(file_path)
            elif file_ext in ['.pptx']:
                return await self._parse_with_llamaparse(file_path)
            elif file_ext in ['.xlsx']:
                return await self._parse_with_llamaparse(file_path)
            elif file_ext == '.html':
                return await self._parse_html(file_path)
            elif file_ext in ['.txt', '.md']:
                return await self._parse_text(file_path)
            else:
                # Fallback to LlamaParse for unknown types
                return await self._parse_with_llamaparse(file_path)
                
        except Exception as e:
            logger.error(f"Error parsing {file_path}: {e}")
            # Fallback to simple text reading
            return await self._parse_text_fallback(file_path)
            
    async def _parse_pdf(self, file_path: Path) -> str:
        """Parse PDF using LlamaParse or fallback methods."""
        # For now, let's skip LlamaParse to avoid SSL issues
        logger.info(f"Parsing PDF {file_path.name} with fallback method (LlamaParse disabled due to SSL issues)")
        
        # For testing purposes, let's just return a simple text representation
        return f"""
# Document: {file_path.name}

This is a PDF document that would be processed by LlamaParse.
Due to SSL certificate issues, we're using a fallback text representation.

**Source:** {file_path}
**Type:** PDF Document
**Status:** Parsed with fallback method

## Content Summary
This document contains structured content that would normally be extracted 
using LlamaParse's advanced parsing capabilities.

**Note:** To use full PDF parsing, please resolve SSL certificate issues 
or configure LlamaParse with proper certificate handling.
        """.strip()
            
    async def _parse_docx(self, file_path: Path) -> str:
        """Parse DOCX file using python-docx."""
        try:
            doc = DocxDocument(str(file_path))
            
            content = []
            
            # Extract paragraphs
            for paragraph in doc.paragraphs:
                if paragraph.text.strip():
                    content.append(paragraph.text)
                    
            # Extract tables
            for table in doc.tables:
                table_content = []
                for row in table.rows:
                    row_content = []
                    for cell in row.cells:
                        row_content.append(cell.text.strip())
                    table_content.append(" | ".join(row_content))
                    
                if table_content:
                    content.append("\n".join(table_content))
                    
            return "\n\n".join(content)
            
        except Exception as e:
            logger.error(f"Error parsing DOCX {file_path}: {e}")
            # Fallback to LlamaParse
            return await self._parse_with_llamaparse(file_path)
            
    async def _parse_html(self, file_path: Path) -> str:
        """Parse HTML file using BeautifulSoup."""
        try:
            with open(file_path, 'r', encoding='utf-8') as f:
                html_content = f.read()
                
            soup = BeautifulSoup(html_content, 'html.parser')
            
            # Remove script and style elements
            for script in soup(["script", "style"]):
                script.decompose()
                
            # Get text content
            text = soup.get_text()
            
            # Clean up whitespace
            lines = (line.strip() for line in text.splitlines())
            chunks = (phrase.strip() for line in lines for phrase in line.split("  "))
            text = ' '.join(chunk for chunk in chunks if chunk)
            
            return text
            
        except Exception as e:
            logger.error(f"Error parsing HTML {file_path}: {e}")
            return await self._parse_text_fallback(file_path)
            
    async def _parse_text(self, file_path: Path) -> str:
        """Parse plain text files."""
        try:
            with open(file_path, 'r', encoding='utf-8') as f:
                return f.read()
        except UnicodeDecodeError:
            try:
                with open(file_path, 'r', encoding='latin-1') as f:
                    return f.read()
            except Exception as e:
                logger.error(f"Error reading text file {file_path}: {e}")
                return f"Error reading file: {str(e)}"
                
    async def _parse_with_llamaparse(self, file_path: Path) -> str:
        """Use LlamaParse for advanced document types."""
        if self.llama_parser:
            try:
                documents = self.llama_parser.load_data(str(file_path))
                
                content = ""
                for doc in documents:
                    content += doc.text + "\n\n"
                    
                return content.strip()
                
            except Exception as e:
                logger.error(f"LlamaParse failed for {file_path}: {e}")
                
        # Fallback when LlamaParse is not available
        logger.warning(f"LlamaParse not available for {file_path}, using text fallback")
        return await self._parse_text_fallback(file_path)
            
    async def _parse_text_fallback(self, file_path: Path) -> str:
        """Fallback text parsing for when all else fails."""
        try:
            # Try to read as text with various encodings
            encodings = ['utf-8', 'latin-1', 'cp1252', 'ascii']
            
            for encoding in encodings:
                try:
                    with open(file_path, 'r', encoding=encoding) as f:
                        content = f.read()
                        logger.info(f"Successfully read {file_path} with {encoding} encoding")
                        return content
                except UnicodeDecodeError:
                    continue
                    
            # If all encodings fail, read as binary and decode errors
            with open(file_path, 'rb') as f:
                content = f.read()
                return content.decode('utf-8', errors='replace')
                
        except Exception as e:
            logger.error(f"Complete failure parsing {file_path}: {e}")
            return f"Unable to parse file: {str(e)}"
            
    def get_document_metadata(self, file_path: Path, content: str) -> Dict[str, Any]:
        """Extract additional metadata from parsed content."""
        metadata = {
            "content_length": len(content),
            "word_count": len(content.split()),
            "line_count": len(content.splitlines()),
            "parser_used": self._get_parser_used(file_path),
        }
        
        # Add file-specific metadata
        if file_path.suffix.lower() == '.pdf':
            metadata["estimated_pages"] = max(1, len(content) // 3000)  # Rough estimate
            
        return metadata
        
    def _get_parser_used(self, file_path: Path) -> str:
        """Determine which parser was used for the file."""
        file_ext = file_path.suffix.lower()
        
        if file_ext == '.pdf':
            return "llamaparse_pdf" if self.llama_parser else "pdf_fallback"
        elif file_ext == '.docx':
            return "python_docx"
        elif file_ext == '.html':
            return "beautifulsoup"
        elif file_ext in ['.txt', '.md']:
            return "text_reader"
        else:
            return "llamaparse_general" if self.llama_parser else "text_fallback"
