requirements-minimal.txt
# Minimal requirements without LlamaParse (for testing)
langgraph==0.1.5
langchain==0.2.5
langchain-google-genai==1.0.6
google-generativeai==0.5.4
neo4j==5.20.0
supabase==2.5.0
postgrest==0.16.5
python-dotenv==1.0.0
pydantic==2.7.4
pydantic-settings==2.3.3
asyncio-mqtt==0.16.1
aiofiles==23.2.1
httpx==0.27.0
rich==13.7.1
typer==0.12.3
pytest==8.2.2
pytest-asyncio==0.23.7
watchdog==4.0.1
python-docx==1.1.0
beautifulsoup4==4.12.3
lxml==5.2.2
urllib3==1.26.18
